requirements.txt
asgiref==3.5.2
certifi==2020.6.20
chardet==4.0.0
class-registry==2.1.2
colorzero==1.1
distlib==0.3.6
distro==1.5.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-nested-routers==0.93.4
environ==1.0
filelock==3.8.0
filters==1.3.2
filters-django==1.0.5